backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
datetime-truncate==1.1.1
Django==5.1
django-admin-charts==1.4.0
django-bower==5.2.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-redis==5.4.0
django-suit==0.2.28
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory_boy==3.3.1
Faker==28.4.1
google-api-core==2.19.2
google-api-python-client==2.145.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
Jinja2==3.1.4
kombu==5.4.1
MarkupSafe==2.1.5
oauth2client==1.5.2
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
redis==5.0.8
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
